ipv6: sr: add Masquerading SR proxy
This patch adds Masquerading SR proxy, called End.AM. End.AM is composed
of two parts: masquerading and de-masquerading. The masquerading part is
implemented as the End.AM function that updates the IPv6 destination
address with SRH[0]. From the configuration viewpoint, and End.AM route
would be configured as follows:

$ ip -6 route add to fc00:1:: encap seg6local action End.AM nh6
fc00:f1::2 table 0 dev eth1

End functions can perform the de-masquerading part of End.AM. The
de-masquerading part is, decrement SL, update IPv6 DA with SRH[SL],
and routing by the updated DA [1]. This is identical to the basic
Endpoint behavior defined in [2].

[1] https://tools.ietf.org/html/draft-xuclad-spring-sr-service-programming-02#section-6.4.1
[2] https://tools.ietf.org/html/draft-ietf-spring-srv6-network-programming-04#section-4.1 (S12 - S15 in the pseudocode of End)

On the other hand, masqueraded IPv6 packets would have arbitrary
destination addresses (the original destinations); thus, de-masquerading
should be appied to ::/0. An End.T on a VRF achieves de-masquerading for
arbitrary masqueraded packets as follows:

$ ip -6 route add vrf vrf-am to default encap seg6local action End.T
table 254 dev eth2

IFACE-IN of the End.AM proxy is enslaved to the vrf-am.

Signed-off-by: Ryo Nakamura <upa@haeena.net>